adminapi: Define exceptions in a central place
